+++ dali/kernels/imgproc/color_manipulation/hsv_cpu.h
@@ -29,8 +29,8 @@ namespace kernels {
 
 namespace hsv {
 
-constexpr size_t kNdims = 3;
-constexpr size_t kNchannels = 3;
+constexpr size_t ndims = 3;
+constexpr size_t nchannels = 3;
 
 /**
  * Defines region of interest.
@@ -45,7 +45,7 @@ constexpr size_t kNchannels = 3;
  *       v  |      +-----+                   |
  *          +--------------------------------+
  */
-template <size_t ndims>
+template<size_t ndims>
 using Roi = Box<ndims, int>;
 
 
@@ -58,84 +58,79 @@ using Roi = Box<ndims, int>;
  * @param nchannels Number of channels in data
  * @return Corresponding TensorShape
  */
-template <size_t ndims_roi>
+template<size_t ndims_roi>
 TensorShape<ndims_roi + 1> roi_shape(Roi<ndims_roi> roi, size_t nchannels) {
-  assert(all_coords(roi.hi >= roi.lo) && "Cannot create a tensor shape from an invalid Box");
-  TensorShape<ndims_roi + 1> ret;
-  auto e = roi.extent();
-  auto ridx = ndims_roi;
-  ret[ridx--] = nchannels;
-  for (size_t idx = 0; idx < ndims_roi; idx++) {
-    ret[ridx--] = e[idx];
-  }
-  return ret;
+    assert(all_coords(roi.hi >= roi.lo) && "Cannot create a tensor shape from an invalid Box");
+    TensorShape<ndims_roi + 1> ret;
+    auto e = roi.extent();
+    auto ridx = ndims_roi;
+    ret[ridx--] = nchannels;
+    for (size_t idx = 0; idx < ndims_roi; idx++) {
+        ret[ridx--] = e[idx];
+    }
+    return ret;
 }
 
 }  // namespace hsv
 
 
-template <class OutputType, class InputType>
+template<class OutputType, class InputType>
 class HsvCpu {
-  // TODO(mszolucha): implement float16
-  static_assert(!std::is_same<OutputType, float16_cpu>::value &&
-                !std::is_same<InputType, float16_cpu>::value, "float16 not implemented yet");
-
  public:
-  using Roi = hsv::Roi<hsv::kNdims - 1>;
+    using Roi = hsv::Roi<hsv::ndims - 1>;
 
 
-  KernelRequirements
-  Setup(KernelContext &context, const InTensorCPU<InputType, hsv::kNdims> &in, float hue,
-        float saturation, float value, const Roi *roi = nullptr) {
-    DALI_ENFORCE(!roi || all_coords(roi->hi >= roi->lo), "Region of interest is invalid");
-    auto adjusted_roi = AdjustRoi(roi, in.shape);
-    KernelRequirements req;
-    TensorListShape<> out_shape({hsv::roi_shape(adjusted_roi, hsv::kNchannels)});
-    req.output_shapes = {std::move(out_shape)};
-    return req;
-  }
-
-
-  void Run(KernelContext &context, const OutTensorCPU<OutputType, hsv::kNdims> &out,
-           const InTensorCPU<InputType, hsv::kNdims> &in, float hue, float saturation, float value,
-           const Roi *roi = nullptr) {
-    auto adjusted_roi = AdjustRoi(roi, in.shape);
-    auto num_channels = in.shape[2];
-    auto image_width = in.shape[1];
-    auto ptr = out.data;
-
-    ptrdiff_t row_stride = image_width * num_channels;
-    auto *row = in.data + adjusted_roi.lo.y * row_stride;
-    for (int y = adjusted_roi.lo.y; y < adjusted_roi.hi.y; y++) {
-      for (int x = adjusted_roi.lo.x; x < adjusted_roi.hi.x; x++) {
-        auto elem = row + x * num_channels;
-        *ptr++ = ConvertSat<OutputType>(*(elem + 0) + hue /*hue hue*/);
-        *ptr++ = ConvertSat<OutputType>(*(elem + 1) * saturation);
-        *ptr++ = ConvertSat<OutputType>(*(elem + 2) * value);
-      }
-      row += row_stride;
+    KernelRequirements
+    Setup(KernelContext &context, const InTensorCPU<InputType, hsv::ndims> &in, float hue,
+          float saturation, float value, const Roi *roi = nullptr) {
+        DALI_ENFORCE(!roi || all_coords(roi->hi >= roi->lo), "Region of interest is invalid");
+        auto adjusted_roi = AdjustRoi(roi, in.shape);
+        KernelRequirements req;
+        TensorListShape<> out_shape({hsv::roi_shape(adjusted_roi, hsv::nchannels)});
+        req.output_shapes = {std::move(out_shape)};
+        return req;
     }
-  }
+
+
+    void Run(KernelContext &context, const OutTensorCPU<OutputType, hsv::ndims> &out,
+             const InTensorCPU<InputType, hsv::ndims> &in, float hue, float saturation, float value,
+             const Roi *roi = nullptr) {
+        auto adjusted_roi = AdjustRoi(roi, in.shape);
+        auto num_channels = in.shape[2];
+        auto image_width = in.shape[1];
+        auto ptr = out.data;
+
+        ptrdiff_t row_stride = image_width * num_channels;
+        auto *row = in.data + adjusted_roi.lo.y * row_stride;
+        for (int y = adjusted_roi.lo.y; y < adjusted_roi.hi.y; y++) {
+            for (int x = adjusted_roi.lo.x; x < adjusted_roi.hi.x; x++) {
+                *ptr++ = ConvertSat<OutputType>(row[x * num_channels + 0] + hue);
+                *ptr++ = ConvertSat<OutputType>(row[x * num_channels + 1] * saturation);
+                *ptr++ = ConvertSat<OutputType>(row[x * num_channels + 2] * value);
+            }
+            row += row_stride;
+        }
+    }
 
 
  private:
-  /**
-   * Adjusted Roi is a Roi, which doesn't overflow the image,
-   * that given by TensorShape.
-   *
-   * In case, when no Roi is provided (roi == nullptr),
-   * size of whole image is returned as Roi.
-   *
-   * Assumes HWC layout
-   */
-  Roi AdjustRoi(const Roi *roi, const TensorShape<hsv::kNdims> &shape) {
-    constexpr size_t spatial_dims = hsv::kNdims - 1;
-    ivec<spatial_dims> size;
-    for (size_t i = 0; i < spatial_dims; i++)
-      size[i] = shape[spatial_dims - 1 - i];
-    Roi whole_image = {0, size};
-    return roi ? intersection(*roi, whole_image) : whole_image;
-  }
+    /**
+     * Adjusted Roi is a Roi, which doesn't overflow the image,
+     * that given by TensorShape.
+     *
+     * In case, when no Roi is provided (roi == nullptr),
+     * size of whole image is returned as Roi.
+     *
+     * Assumes HWC layout
+     */
+    Roi AdjustRoi(const Roi *roi, const TensorShape<hsv::ndims> &shape) {
+        constexpr size_t spatial_dims = hsv::ndims - 1;
+        ivec<spatial_dims> size;
+        for (size_t i = 0; i < spatial_dims; i++)
+            size[i] = shape[spatial_dims - 1 - i];
+        Roi whole_image = {0, size};
+        return roi ? intersection(*roi, whole_image) : whole_image;
+    }
 };
 
 }  // namespace kernels
